// src/views/risk_control/playerPortraitMonitoring_management/utils/types.ts
interface TagScoreItem {
  tagID?: number;
  score?: number;
}